refactor: centralize filter and timer defaults
