+++ src/constants/timers.ts
@@ -1,0 +1,22 @@
+export const TIMER_LABELS: Record<string, string> = {
+  "cetus-cycle":    "Cetus",
+  "vallis-cycle":   "Orb Vallis",
+  "cambion-cycle":  "Cambion Drift",
+  "zariman-cycle":  "Zariman",
+  "bounty-cetus":   "Cetus Bounties",
+  "bounty-vallis":  "Vallis Bounties",
+  "bounty-cambion": "Cambion Bounties",
+  "bounty-zariman": "Zariman Bounties",
+  "bounty-hex":     "Hex Bounties",
+  "sortie":         "Sortie",
+  "archon-hunt":    "Archon Hunt",
+  "daily-reset":    "Daily Reset",
+  "weekly-reset":   "Weekly Reset",
+  "void-trader":    "Void Trader",
+  "nightwave":        "Nightwave",
+  "prime-resurgence": "Prime Resurgence",
+  "circuit":        "The Circuit",
+  "duviri-spiral":  "Duviri Spiral",
+  "kahl":           "Kahl / Break Narmer",
+  "deep-archimedea":"Deep Archimedea",
+};
